requirements.txt
alpaca-py==0.42.2
anyio==4.11.0
ccxt==4.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coinbase-advanced-py==1.8.2
colorama==0.4.6
coverage==7.10.7
distlib==0.4.0
eventkit==1.0.3
fastapi==0.119.0
feedparser==6.0.12
filelock==3.19.1
fsspec==2025.9.0
git-filter-repo==2.47.0
h11==0.16.0
httptools==0.7.1
huggingface-hub==0.35.1
ib-insync==0.9.86
ibapi==9.81.1.post1
idna==3.11
markupsafe==3.0.2
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
packaging==25.0
platformdirs==4.5.0
propcache==0.4.1
pyarrow==21.0.0
pyjwt==2.10.1
pytest==8.4.1
pytest-cov==5.0.0
python-dotenv==1.1.1
pyyaml==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sgmllib3k==1.0.0
sniffio==1.3.1
sseclient-py==1.8.0
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
tree-sitter==0.24.0
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.23.2
tree-sitter-language-pack==0.9.0
tree-sitter-yaml==0.7.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
vaderSentiment==3.3.2
virtualenv==20.34.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wheel==0.45.1
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0


